+++ client/web-sveltekit/src/lib/repo/FileDiffHunks.svelte
@@ -68,8 +68,6 @@
     }
 
     td {
-        background-color: var(--code-bg);
-
         &.num {
             min-width: 2.5rem;
             line-height: 1.6666666667;