+++ client/web-sveltekit/src/routes/[...repo=reporev]/(validrev)/(code)/+layout.svelte
@@ -183,7 +183,8 @@
                             history={commitHistory}
                             loading={$commitHistoryQuery?.fetching ?? true}
                             fetchMore={commitHistoryQuery.fetchMore}
-                            enableInlineDiffs={$page.route.id?.includes('/blob/') ?? false}
+                            enableInlineDiff={$page.data.enableInlineDiff}
+                            enableViewAtCommit={$page.data.enableViewAtCommit}
                         />
                     {/key}
                 </TabPanel>
@@ -208,7 +209,6 @@
     section {
         display: flex;
         flex: 1;
-        background-color: var(--code-bg);
         overflow: hidden;
     }
 
@@ -266,8 +266,8 @@
         overflow: hidden;
         border-top: 1px solid var(--border-color);
         box-shadow: var(--bottom-panel-shadow);
-        background-color: var(--code-bg);
-        padding: 0 0.25rem;
+        background-color: var(--color-bg-1);
+        color: var(--text-body);
 
         &.open {
             height: 32vh;