+++ client/web-sveltekit/src/routes/search/FileContentSearchResult.svelte
@@ -135,7 +135,7 @@
         text-align: left;
         border: none;
         padding: 0.25rem 0.5rem;
-        background-color: var(--code-bg);
+        background-color: var(--body-bg);
         color: var(--text-muted);
         cursor: pointer;
 
@@ -145,7 +145,7 @@
         }
 
         &:hover {
-            background-color: var(--subtle-bg-2);
+            background-color: var(--color-bg-2);
         }
     }