+++ internal/appliance/blobstore.go
@@ -35,20 +35,11 @@ func (r *Reconciler) reconcileBlobstore(ctx context.Context, sg *Sourcegraph, ow
 
 func buildBlobstorePersistentVolumeClaim(sg *Sourcegraph) (corev1.PersistentVolumeClaim, error) {
 	storage := sg.Spec.Blobstore.StorageSize
-	if storage == "" {
-		storage = "100Gi"
-	}
-
 	if _, err := resource.ParseQuantity(storage); err != nil {
 		return corev1.PersistentVolumeClaim{}, errors.Errorf("invalid blobstore storage size: %s", storage)
 	}
 
-	storageClassName := sg.Spec.StorageClass.Name
-	if storageClassName == "" {
-		storageClassName = "sourcegraph"
-	}
-
-	p := pvc.NewPersistentVolumeClaim("blobstore", sg.Namespace, resource.MustParse(storage), storageClassName)
+	p := pvc.NewPersistentVolumeClaim("blobstore", sg.Namespace, resource.MustParse(storage), sg.Spec.StorageClass.Name)
 
 	return p, nil
 }